+++ src/cljs-lib/src/calva/repl/webview/core.cljs
@@ -135,14 +135,30 @@
   (.. vscode -window
       (onDidChangeActiveColorTheme
        (fn [e]
+         ;; Ensure syntax highlighting is not broken after the output view is moved into or out of the main
+         ;; VS Code window. See https://github.com/BetterThanTomorrow/calva/issues/2895.
+         ;; This is probably called more frequently than it needs to be, but if that's problematic we can try to
+         ;; figure out how to identify when this event means the output view moved into or out of the main window.
+         ;; This event fires for more cases than just that, such as when it's focused or unfocused.
          (set-code-theme! context {:color-theme-kind (.. e -kind)
                                    :webview-panel webview-panel})))))
+
+(defn create-view-state-change-listener
+  [{:keys [^js vscode/vscode] :as context}
+   {:keys [^js webview-panel]}]
+  (.. webview-panel
+      (onDidChangeViewState
+       (fn [^js _event]
+         (set-code-theme! context
+                          {:color-theme-kind (.. ^js vscode -window -activeColorTheme -kind)
+                           :webview-panel webview-panel})))))
 
 (defn add-subscriptions!
   [{vscode-context :vscode/context
     :as context}
    {:keys [webview-panel]}]
-  (let [subscriptions [(create-color-theme-change-listener context {:webview-panel webview-panel})]]
+  (let [subscriptions [(create-color-theme-change-listener context {:webview-panel webview-panel})
+                       (create-view-state-change-listener context {:webview-panel webview-panel})]]
     (run! (fn [subscription]
             (.. ^js vscode-context -subscriptions (push subscription)))
           subscriptions)))
@@ -173,6 +189,10 @@
                                 ;; See also: https://code.visualstudio.com/api/references/vscode-api#WebviewPanelOptions
                                 ;; "retainContextWhenHidden has a high memory overhead and should only be used if your
                                 ;; panel's context cannot be quickly saved and restored."
+                                ;; Content reloading using setState and getState and message passing between the webview
+                                ;; and the extension was attempted, but it proved to be troublesome, so it was removed.
+                                ;; If someone wants to attempt to add it again, here's the PR for the removal:
+                                ;; https://github.com/BetterThanTomorrow/calva/pull/2896
                                 :retainContextWhenHidden true
                                 :enableFindWidget true}))]
     (initialize-webview-panel context webview-panel)